Rename: "components 폴더를 view 폴더 안으로 이동"
