+++ static/js/view/components/footer.js
