+++ static/js/view/components/mainBanner.js
