+++ static/js/view/components/promotionBanner.js
