+++ static/js/view/components/sectionBasic.js
