+++ static/js/view/components/snb.js
